+++ packages/expo-dev-launcher/android/src/debug/java/expo/modules/devlauncher/modules/DevLauncherInternalModule.kt
@@ -262,13 +262,11 @@ class DevLauncherInternalModule(reactContext: ReactApplicationContext?) :
   }
 
   private fun getApplicationIconUri(): String {
-    var appIcon = ""
     val packageManager = reactApplicationContext.packageManager
     val packageName = reactApplicationContext.packageName
     val applicationInfo = packageManager.getApplicationInfo(packageName, 0)
-    appIcon = "" + applicationInfo.icon
 //    TODO - figure out how to get resId for AdaptiveIconDrawable icons
-    return appIcon
+    return "" + applicationInfo.icon
   }
 
   @ReactMethod