+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/jni/JavaScriptModuleObject.kt
@@ -1,6 +1,7 @@
 package expo.modules.kotlin.jni
 
 import com.facebook.jni.HybridData
+import com.facebook.react.bridge.NativeMap
 import expo.modules.core.interfaces.DoNotStrip
 
 /**
@@ -18,6 +19,11 @@ class JavaScriptModuleObject {
   private val mHybridData = initHybrid()
 
   private external fun initHybrid(): HybridData
+
+  /**
+   * Exports constants
+   */
+  external fun exportConstants(constants: NativeMap)
 
   /**
    * Register a promise-less function on the CPP module representation.