+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/service/delegates/ExpoHandlingDelegate.kt
@@ -60,7 +60,7 @@ class ExpoHandlingDelegate(protected val context: Context) : HandlingDelegate {
      *   - the foreground main Activity
      *   - the background [NotificationForwarderActivity] Activity that send notification clicked events through broadcast
      */
-    fun createPendingIntentForOpeningApp(context: Context, broadcastIntent: Intent): PendingIntent {
+    fun createPendingIntentForOpeningApp(context: Context, broadcastIntent: Intent, notificationResponse: NotificationResponse): PendingIntent {
       var intentFlags = PendingIntent.FLAG_UPDATE_CURRENT
       if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
         intentFlags = intentFlags or PendingIntent.FLAG_IMMUTABLE
@@ -70,6 +70,7 @@ class ExpoHandlingDelegate(protected val context: Context) : HandlingDelegate {
         Intent()
       }
       foregroundActivityIntent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_MULTIPLE_TASK
+      NotificationsService.setNotificationResponseToIntent(foregroundActivityIntent, notificationResponse)
       val backgroundActivityIntent = Intent(context, NotificationForwarderActivity::class.java)
       backgroundActivityIntent.putExtras(broadcastIntent)
       return PendingIntent.getActivities(context, 0, arrayOf(foregroundActivityIntent, backgroundActivityIntent), intentFlags)
@@ -99,9 +100,18 @@ class ExpoHandlingDelegate(protected val context: Context) : HandlingDelegate {
       getListeners().forEach {
         it.onNotificationReceived(notification)
       }
-    } else {
+    } else if (notification.shouldPresent()) {
       NotificationsService.present(context, notification)
     }
+  }
+
+  /**
+   * If the app is backgrounded, a notification is only presented if
+   * the title and or text is present. If both are null or empty, this is a "data-only" or "silent"
+   * notification that should not be presented to the user.
+   */
+  private fun Notification.shouldPresent(): Boolean {
+    return !(notificationRequest.content.title.isNullOrEmpty() && notificationRequest.content.text.isNullOrEmpty())
   }
 
   override fun handleNotificationResponse(notificationResponse: NotificationResponse) {